requirements.txt
# Core
torch>=2.1.0
stable-baselines3>=2.3.0
gymnasium>=0.29.0
pettingzoo>=1.24.0

# Experiment tracking
wandb>=0.17.0

# Numerical
numpy>=1.26.0
scipy>=1.12.0

# Visualization
matplotlib>=3.8.0
pygame>=2.5.0

# Config
pyyaml>=6.0
hydra-core>=1.3.0

# Agent automation
PyGithub>=2.3.0
openai>=1.30.0

# Dev tools
ipykernel>=6.29.0
jupyter>=1.0.0
black>=24.0.0
ruff>=0.4.0
pytest>=8.0.0
